fix: crc64

+++ src/kv/storage/crc64.rs
@@ -1,4 +1,4 @@
-const crctab64: [usize; 256] = [
+const crctab64: [u64; 256] = [
     0x0000000000000000, 0x7ad870c830358979, 0xf5b0e190606b12f2,
     0x8f689158505e9b8b, 0xc038e5739841b68f, 0xbae095bba8743ff6,
     0x358804e3f82aa47d, 0x4f50742bc81f2d04, 0xab28ecb46814fe75,
@@ -87,10 +87,10 @@ const crctab64: [usize; 256] = [
     0x29b7d047efec8728
 ];
 
-pub fn kv_crc64(buffer: &Vec<u8>) -> usize {
-    let mut crc: usize = 0;
+pub fn kv_crc64(buffer: &Vec<u8>) -> u64 {
+    let mut crc: u64 = 0;
     for x in buffer {
-        crc = crctab64[(crc ^ x.clone() as usize) & 0xff] ^ (crc >> 8);
+        crc = crctab64[((crc ^ x.clone() as u64) & 0xff) as usize] ^ (crc >> 8);
     }
     return crc;
 }
